post probleme fix + commentaire fix
on peut désormais poster + d'1 commentaire par personne

+++ typescript-mysql-quickstart/source/controllers/vecoleo/probleme/commentaire.ts
@@ -21,90 +21,95 @@ const NAMESPACE = 'commentaire';
 const createCommentaire = async (req: Request, res: Response, next: NextFunction) => {  //TODO
     logging.info(NAMESPACE, 'Inserting commentaire');
     console.log("mon body = ", req.body)    //DEBUG
-    if(!req.body || !req.body.probleme_id || !req.body.utilisateur_id || !req.body.description){
+    if (!req.body || !req.body.probleme_id || !req.body.utilisateur_id || !req.body.description) {
         res.status(400);
         res.send("le body ne contiens pas d'information pour le create, veuillez ajouter le json contenant les donnés dans le body.");
         return;
     }
-    const probleme_id : number = parseInt(req.body.probleme_id);
-    const utilisateur_id : number = parseInt(req.body.utilisateur_id);
+    const probleme_id: number = parseInt(req.body.probleme_id);
+    const utilisateur_id: number = parseInt(req.body.utilisateur_id);
     let resume: string
     if (req.body.resume) {
-       resume = req.body.resume
+        resume = req.body.resume
     }
     else {
         resume = ""
     }
-    
-    const description:string = req.body.description;
+
+    const description: string = req.body.description;
 
     let query = `INSERT INTO commentaire (probleme_id, utilisateur_id, resume, description) VALUES (${probleme_id}, ${utilisateur_id}, "${resume}", "${description}")`;
-    
+
     return await executeSQLCommand(req, res, next, NAMESPACE, query, 'commentaire created: ');
 };
 
 const getAllCommentaire = async (req: Request, res: Response, next: NextFunction) => { //TODO
     logging.info(NAMESPACE, 'Getting all commentaire.');
     const query = `SELECT commentaire.* FROM commentaire`;
-    
+
     return await executeSQLCommand(req, res, next, NAMESPACE, query, 'Retrieved commentaire list : ');
 };
 
 const getAllCommentaireFromUser = async (req: Request, res: Response, next: NextFunction) => { //TODO
     logging.info(NAMESPACE, 'Getting all commentaire by utilisateur_id.');
-    if(!req.params.utilisateur_id){
+    if (!req.params.utilisateur_id) {
         res.status(400);
         res.send("erreur, les arguments doivent être l'id du commentaire");
         return;
     }
     const query = `SELECT commentaire.* FROM commentaire WHERE commentaire.utilisateur_id = ${req.params.utilisateur_id}`;
-    
+
     return await executeSQLCommand(req, res, next, NAMESPACE, query, 'Retrieved commentaire list : ');
 };
 
 const getAllCommentaireFromProbleme = async (req: Request, res: Response, next: NextFunction) => { //TODO
     logging.info(NAMESPACE, 'Getting all commentaire by utilisateur_id.');
-    if(!req.params.probleme_id){
+    if (!req.params.probleme_id) {
         res.status(400);
         res.send("erreur, les arguments doivent être l'id du commentaire");
         return;
     }
     const query = `SELECT commentaire.* FROM commentaire WHERE commentaire.probleme_id = ${req.params.probleme_id}`;
-    
+
     return await executeSQLCommand(req, res, next, NAMESPACE, query, 'Retrieved image list : ');
 };
 
 const updateOneCommentaireById = async (req: Request, res: Response, next: NextFunction) => {//TODO
-    if(!req.params || !req.body.commentaire_id || (!req.body.resume && !req.body.description)){
+    if (!req.params || !req.body.commentaire_id || (!req.body.resume && !req.body.description)) {
         res.status(400);
         res.send("le body ne contiens pas d'information pour le update, veuillez ajouter le json contenant les donnés dans le body ou le .");
         return;
     }
-    const commentaire_id : number = parseInt(req.body.commentaire_id);
+    const commentaire_id: number = parseInt(req.body.commentaire_id);
     const resume: string = req.body.resume;
-    const description:string = req.body.description;
-    
+    const description: string = req.body.description;
+
     let query = `UPDATE commentaire SET `
-    if(resume){
+    if (resume) {
         query += `adresse = \"${resume}\" ,`
     }
-    if(description){
+    if (description) {
         query += `description = "${description}"  `
     }
     query = query.substring(0, query.length - 1)
     query += `WHERE commentaire.commentaire_id = ${commentaire_id}`
-    
+
     return await executeSQLCommand(req, res, next, NAMESPACE, query, "updating problemes: ");
 };
 
 const DeleteOneCommentaireById = async (req: Request, res: Response, next: NextFunction) => {//TODO
     logging.info(NAMESPACE, 'DELETE one commentaire by id.');
-    if(!req.params || !req.params.commentaire_id){
+    if (!req.params || !req.params.commentaire_id) {
         res.status(400);
         res.send("le body ne contiens pas d'information pour le DELETE, veuillez ajouter le json contenant les donnés dans le body.");
         return;
     }
-    const commentaire_id : number = parseInt(req.body.commentaire_id);
+    const commentaire_id: number = parseInt(req.params.commentaire_id);
+    if (isNaN(commentaire_id)) {
+        res.status(400);
+        res.send("Le champ commentaire_id doit être un nombre valide.");
+        return;
+    }
     const query = `DELETE FROM commentaire WHERE commentaire.commentaire_id = ${commentaire_id}`;
 
     return await executeSQLCommand(req, res, next, NAMESPACE, query, 'delete commentaire : ');
@@ -112,12 +117,12 @@ const DeleteOneCommentaireById = async (req: Request, res: Response, next: NextF
 
 const DeleteOneCommentaireFromUser = async (req: Request, res: Response, next: NextFunction) => {//TODO
     logging.info(NAMESPACE, 'DELETE all commentaire from userId.');
-    if(!req.params || !req.params.utilisateur_id){
+    if (!req.params || !req.params.utilisateur_id) {
         res.status(400);
         res.send("le body ne contiens pas d'information pour le DELETE, veuillez ajouter le json contenant les donnés dans le body.");
         return;
     }
-    const utilisateur_id : number = parseInt(req.params.utilisateur_id);
+    const utilisateur_id: number = parseInt(req.params.utilisateur_id);
     const query = `DELETE FROM commentaire WHERE commentaire.utilisateur_id = ${utilisateur_id}`;
 
     return await executeSQLCommand(req, res, next, NAMESPACE, query, 'delete commentaire : ');
@@ -125,24 +130,24 @@ const DeleteOneCommentaireFromUser = async (req: Request, res: Response, next: N
 
 const DeleteOneCommentaireFromProbleme = async (req: Request, res: Response, next: NextFunction) => {//TODO
     logging.info(NAMESPACE, 'DELETE all commentaire from probleme Id.');
-    if(!req.params || !req.params.probleme_id){
+    if (!req.params || !req.params.probleme_id) {
         res.status(400);
         res.send("le body ne contiens pas d'information pour le DELETE, veuillez ajouter le json contenant les donnés dans le body.");
         return;
     }
-    const probleme_id : number = parseInt(req.params.probleme_id);
+    const probleme_id: number = parseInt(req.params.probleme_id);
     const query = `DELETE FROM commentaire WHERE commentaire.probleme_id = ${probleme_id}`;
 
     return await executeSQLCommand(req, res, next, NAMESPACE, query, 'delete commentaire : ');
 };
 
 export default {
-    createCommentaire, 
-    getAllCommentaire, 
-    getAllCommentaireFromUser, 
-    getAllCommentaireFromProbleme, 
-    updateOneCommentaireById, 
-    DeleteOneCommentaireById, 
-    DeleteOneCommentaireFromUser, 
+    createCommentaire,
+    getAllCommentaire,
+    getAllCommentaireFromUser,
+    getAllCommentaireFromProbleme,
+    updateOneCommentaireById,
+    DeleteOneCommentaireById,
+    DeleteOneCommentaireFromUser,
     DeleteOneCommentaireFromProbleme
 };
